requirements.txt
flask
python-telegram-bot
pydub
whisper
torch
torchvision
torchaudio